Add pre-computed results, roadmap, and fix two query errors
- results/ (12 files): captured output from all analysis and internals scripts
- ROADMAP.md: future improvement ideas (dbt, CI, streaming, ML)
- Fix ROUND(double precision) cast in 02_revenue.sql Pareto query
- Fix IMMUTABLE expression index in 04_optimization.sql (use plain B-tree)
- README: add Pre-computed Results section

+++ analysis/02_revenue.sql
@@ -105,7 +105,7 @@ SELECT
     END                         AS lead_time_bucket,
     COUNT(*)                    AS tickets,
     ROUND(AVG(amount), 2)       AS avg_price_rub,
-    ROUND(PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY amount), 2) AS median_price_rub
+    ROUND((PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY amount))::numeric, 2) AS median_price_rub
 FROM booking_lead
 GROUP BY lead_time_bucket
 ORDER BY lead_time_bucket;
